requirements.txt
anthropic
python-dotenv
pyautogui
keyboard